Use the column part of URIs (first off, vim only)
This implements the usage  of the optional `column` element of open-in-
editor URIs if vim (or a derivative) is the configured editor.

This feature shouldn't be hard to add to the other classes once we have
figured out these editors' according command line syntaxes.

Btw, I haven't missed to also adjust the `script` file accordingly this
time. ;)