Log at error and info level by default (#20)
It's very annoying to debug this since it's not running in a terminal,
and it involves obscure OS mechanisms that hardly anyone's familiar
with. Having the info-level logging output present by default makes it
easy to determine, when it's not working, whether the script was ever
invoked by the URL handler.